+++ backend/src/main/java/org/cathori/backend/notification/infra/PushNotificationScheduler.java
@@ -1,7 +1,6 @@
 package org.cathori.backend.notification.infra;
 
 import org.cathori.backend.notification.application.push.PushNotificationService;
-import org.cathori.backend.notification.application.push.RePushNotificationService;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
 
@@ -20,7 +19,7 @@ import lombok.extern.slf4j.Slf4j;
 public class PushNotificationScheduler {
 
     private final PushNotificationService pushNotificationService;
-    private final RePushNotificationService retryPushNotificationService;
+
     /**
      * 새 공지 알림을 일괄 발송합니다. (기본: 매일 오전 11시)
      *
@@ -36,7 +35,7 @@ public class PushNotificationScheduler {
     /** 20분마다 실패한 알림 발송을 재시도합니다. */
     @Scheduled(cron = "0 0/20 * * * *")
     public void retryFailedAlerts() {
-        retryPushNotificationService.retryFailedAlerts();
+        pushNotificationService.retryFailedAlerts();
     }
 
 }
